Split the app from starter
