+++ src/App.js
@@ -1,5 +1,5 @@
-import Navbar from './components/Navbar';
-import Main from './components/Main';
+import Navbar from './components/Navbar/Navbar';
+import Main from './components/Main/Main';
 
 export default function App() {
   return (
