+++ src/components/Main/ListBox.jsx
@@ -1,0 +1,19 @@
+import { useState } from 'react';
+import MovieList from './MovieList';
+
+export default function ListBox() {
+  const [isOpen1, setIsOpen1] = useState(true);
+
+  return (
+    <div className="box">
+      <button
+        className="btn-toggle"
+        onClick={() => setIsOpen1((open) => !open)}
+        type="button"
+      >
+        {isOpen1 ? '–' : '+'}
+      </button>
+      {isOpen1 && <MovieList />}
+    </div>
+  );
+}
