+++ src/components/Main/Main.jsx
@@ -1,0 +1,11 @@
+import ListBox from './ListBox';
+import WatchedBox from './WatchedBox';
+
+export default function Main() {
+  return (
+    <main className="main">
+      <ListBox />
+      <WatchedBox />
+    </main>
+  );
+}
