+++ src/components/Main/WatchedBox.jsx
@@ -1,0 +1,27 @@
+import { useState } from 'react';
+import { tempWatchedData } from '../../data/tempMovieData';
+import WatchedSummary from './WatchedSummary';
+import WatchedMovieList from './WatchedMovieList';
+
+export default function WatchedBox() {
+  const [watched, setWatched] = useState(tempWatchedData);
+  const [isOpen2, setIsOpen2] = useState(true);
+
+  return (
+    <div className="box">
+      <button
+        className="btn-toggle"
+        onClick={() => setIsOpen2((open) => !open)}
+        type="button"
+      >
+        {isOpen2 ? '–' : '+'}
+      </button>
+      {isOpen2 && (
+        <>
+          <WatchedSummary watched={watched} />
+          <WatchedMovieList watched={watched} />
+        </>
+      )}
+    </div>
+  );
+}
