+++ src/components/Main/WatchedMovieList.jsx
@@ -1,0 +1,11 @@
+import WatchedMovie from './WatchedMovei';
+
+export default function WatchedMovieList({ watched }) {
+  return (
+    <ul className="list">
+      {watched.map((movie) => (
+        <WatchedMovie movie={movie} key={movie.imdbID} />
+      ))}
+    </ul>
+  );
+}
